70% speed up by reusing buffers for bufio.Reader/Writer
It was really slow to recreate a new pair of bufio.Reader and
bufio.Writer because we had to allocate new memory every time, which
gave the garbage collector a lot of work I'm guessing. I used a pool
to reuse those readers/writers (and hence underlying buffers), and it's
so much faster and it allocates much less memory, it's crazy.