cli: if no args: read from stdin or print help
That means that the program doesn't depend as much on our ability to
determine whether or not we have stuff in stdin.

The reason I make this improvement is because on python 3.3, on windows,
if you start jsoncomma with subprocess, it thinks there's some stuff on
stdin whatever you (sanely) do. jsoncomma/sublime-jsoncomma#10

Here, it'll detect that I'm starting the server subcommand and hence
don't bother checking from stdin. It's not as user friendly because
someone could do

echo "[1, 2,]" | jsoncomma server

And somehow expect the json to be fixed. They wouldn't get any error
letting them know that stdin is completely ignored