add /shutdown endpoint
this is required by clients, like Sublime Text, who don't have an
on_exit event on which we can kill the server.

It writes a json object {"timedout": <bool>}. If true, then that means
that the server has aborted some handlers.

This solution is not theoretically correct, because I don't know how to
properly wait for all handlers to be done in Go (see comments in
main.go)