allow commas when consuming space/comments
We always assumed that were were no commas in the first place, and
because they don't count as space, then the parser would think that 1,2
didn't need a comma between them (ie there was nothing between 1 and 2,
so of course it would write 12). And because we don't write commas by
default, of course we get `12`. There was a test for that which I'm
pretty sure never passed. But it went unnoticed because it was always
disabled. Go me...